+++ czkawka_cli/src/main.rs
@@ -329,6 +329,8 @@ fn similar_videos(similar_videos: SimilarVideosArgs, stop_flag: &Arc<AtomicBool>
         skip_forward_amount,
         scan_duration,
         crop_detect,
+        false, // creating thumbnails in CLI, makes almost no sense
+        10,    // creating thumbnails in CLI, makes almost no sense
     );
     let mut tool = SimilarVideos::new(params);