+++ czkawka_cli/src/progress.rs
@@ -65,6 +65,7 @@ pub(crate) fn get_progress_message(progress_data: &ProgressData) -> String {
         CurrentStage::SimilarImagesCalculatingHashes => "Calculating image hashes",
         CurrentStage::SimilarImagesComparingHashes => "Comparing image hashes",
         CurrentStage::SimilarVideosCalculatingHashes => "Reading similar values",
+        CurrentStage::SimilarVideosCreatingThumbnails => "Creating video thumbnails",
         CurrentStage::BrokenFilesChecking => "Checking broken files",
         CurrentStage::BadExtensionsChecking => "Checking extensions of files",
         CurrentStage::DeletingFiles => "Deleting files/folders",