+++ czkawka_core/src/tools/similar_videos/traits.rs
@@ -12,6 +12,7 @@ use crate::common::progress_data::ProgressData;
 use crate::common::tool_data::{CommonData, CommonToolData, DeleteMethod};
 use crate::common::traits::{AllTraits, DebugPrint, DeletingItems, PrintResults, Search};
 use crate::flc;
+use crate::tools::similar_videos::core::{format_bitrate_opt, format_duration_opt};
 use crate::tools::similar_videos::{Info, SimilarVideos, SimilarVideosParameters};
 
 impl AllTraits for SimilarVideos {}
@@ -78,25 +79,57 @@ impl DebugPrint for SimilarVideos {
 
 impl PrintResults for SimilarVideos {
     fn write_results<T: Write>(&self, writer: &mut T) -> std::io::Result<()> {
+        fn write_video_entry<T: Write>(writer: &mut T, file_entry: &crate::tools::similar_videos::VideosEntry) -> std::io::Result<()> {
+            let bitrate = format_bitrate_opt(file_entry.bitrate);
+            let fps = file_entry.fps.map(|e| format!("{e:.2}")).unwrap_or_default();
+            let codec = file_entry.codec.clone().unwrap_or_default();
+            let dimensions = if let (Some(w), Some(h)) = (file_entry.width, file_entry.height) {
+                format!("{w}x{h}")
+            } else {
+                "".to_string()
+            };
+            let duration = format_duration_opt(file_entry.duration);
+
+            writeln!(
+                writer,
+                "\"{}\" - {} - {} - {} - {} - {} - {}",
+                file_entry.path.to_string_lossy(),
+                format_size(file_entry.size, BINARY),
+                bitrate,
+                fps,
+                codec,
+                dimensions,
+                duration
+            )
+        }
+
         if !self.similar_vectors.is_empty() {
             write!(writer, "{} videos which have similar friends\n\n", self.similar_vectors.len())?;
 
             for struct_similar in &self.similar_vectors {
-                writeln!(writer, "Found {} videos which have similar friends", struct_similar.len())?;
+                writeln!(
+                    writer,
+                    "Found {} videos which have similar friends (path, size, bitrate, fps, codec, dimensions, duration)",
+                    struct_similar.len()
+                )?;
                 for file_entry in struct_similar {
-                    writeln!(writer, "\"{}\" - {}", file_entry.path.to_string_lossy(), format_size(file_entry.size, BINARY))?;
+                    write_video_entry(writer, file_entry)?;
                 }
                 writeln!(writer)?;
             }
         } else if !self.similar_referenced_vectors.is_empty() {
-            write!(writer, "{} videos which have similar friends\n\n", self.similar_referenced_vectors.len())?;
+            write!(
+                writer,
+                "{} videos which have similar friends (path, size, bitrate, fps, codec, dimensions, duration)\n\n",
+                self.similar_referenced_vectors.len()
+            )?;
 
             for (fe, struct_similar) in &self.similar_referenced_vectors {
                 writeln!(writer, "Found {} videos which have similar friends", struct_similar.len())?;
                 writeln!(writer)?;
-                writeln!(writer, "\"{}\" - {}", fe.path.to_string_lossy(), format_size(fe.size, BINARY))?;
+                write_video_entry(writer, fe)?;
                 for file_entry in struct_similar {
-                    writeln!(writer, "\"{}\" - {}", file_entry.path.to_string_lossy(), format_size(file_entry.size, BINARY))?;
+                    write_video_entry(writer, file_entry)?;
                 }
                 writeln!(writer)?;
             }