+++ czkawka_gui/src/connect_things/connect_button_search.rs
@@ -635,6 +635,8 @@ fn similar_video_search(
                 DEFAULT_SKIP_FORWARD_AMOUNT,
                 DEFAULT_VID_HASH_DURATION,
                 DEFAULT_CROP_DETECT,
+                false, // Not implemented in gtk gui
+                10,    // Not implemented in gtk gui
             );
             let mut tool = SimilarVideos::new(params);